Make rendered triangle rotate

+++ src/main/scala/FixedTimestep.scala
@@ -1,7 +1,7 @@
 package games.fantasy.lwjgl
 
-case class FixedTimeStep(accumulator: Double = 0, alpha: Double = 0) {
-  def addDeltaToAccumulator(delta: Double): FixedTimeStep = this.copy(accumulator + delta, alpha)
-  def subtractIntervalFromAccumulator(interval: Double): FixedTimeStep = this.copy(accumulator - interval, alpha)
-  def updateAlpha(interval: Double): FixedTimeStep = this.copy(accumulator, accumulator / interval)
+case class FixedTimeStep(accumulator: Float = 0f, alpha: Float = 0f) {
+  def addDeltaToAccumulator(delta: Float): FixedTimeStep = this.copy(accumulator + delta, alpha)
+  def subtractIntervalFromAccumulator(interval: Float): FixedTimeStep = this.copy(accumulator - interval, alpha)
+  def updateAlpha(interval: Float): FixedTimeStep = this.copy(accumulator, accumulator / interval)
 }
